`Observations::LocationsController`
Combines two controllers I created without really understanding what they did: `Locations::MergesController` (a GET form action) and `Locations::ObservationsController` (a PUT/PATCH callback for the same form) into one controller.

The form gives a UI to deal with “Observations that have a `where` but no `Location`". In other words, they just have a string a user entered, that maybe matches one or more of our Locations. So it’s not a merge, properly speaking. It’s a bulk assignment of a Location association, where there was no association before. It seems to belong more properly under "Observations".

Clears up confusion from my misleadingly-named `Locations::MergesController`: this form does not merge Location records.
