--- app/views/controllers/locations/merges/_row.html.erb
@@ -1,9 +1,0 @@
-<tr>
-  <td>
-    <%= patch_button(name: location.display_name.t,
-                    path: add_observation_to_location_path(
-                      where: @where, location: location
-                    ),
-                    data: { confirm: :list_merge_are_you_sure.l }) %>
-  </td>
-</tr>
