+++ app/views/controllers/observations/locations/_row.html.erb
@@ -1,0 +1,9 @@
+<tr>
+  <td>
+    <%= patch_button(name: location.display_name.t,
+                     path: assign_location_to_observations_path(
+                       where: @where, location: location
+                     ),
+                     data: { confirm: :list_merge_are_you_sure.l }) %>
+  </td>
+</tr>
