+++ app/views/controllers/observations/locations/edit.html.erb
@@ -2,7 +2,8 @@
 @container = :wide
 add_page_title(:list_merge_options_title.t(where: @where))
 # NOTE: the "forms" are patch buttons in the "row" partial
-# TODO: Redo partial, passing a collection?
+# TODO: Redo partial, passing a collection? Or use helper?
+# TODO: Use list group, not table
 %>
 
 <div class="row">
@@ -14,7 +15,7 @@ add_page_title(:list_merge_options_title.t(where: @where))
       </div>
       <table class="table table-striped table-location-merge-options">
         <% @matches.each do |location| %>
-          <%= render(partial: "locations/merges/row",
+          <%= render(partial: "observations/locations/row",
                      locals: { location: location }) %>
         <% end %>
       </table>
@@ -28,7 +29,7 @@ add_page_title(:list_merge_options_title.t(where: @where))
       </div>
       <table class="table table-striped table-location-merge-options">
         <% @others.each do |location| %>
-          <%= render(partial: "locations/merges/row",
+          <%= render(partial: "observations/locations/row",
                      locals: { location: location }) %>
         <% end %>
       </table>
